fix: move signaturetest to internal package
