+++ signature/cose/envelope_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/fxamacker/cbor/v2"
 	"github.com/notaryproject/notation-core-go/signature"
-	"github.com/notaryproject/notation-core-go/signature/signaturetest"
+	"github.com/notaryproject/notation-core-go/signature/internal/signaturetest"
 	"github.com/notaryproject/notation-core-go/testhelper"
 	"github.com/veraison/go-cose"
 )
